fix(frontend): exclude modals-demo route from production builds (#1630)
Gate the modal playground behind a production notFound() guard and load
demo content via a dev-only dynamic import so the route and its bundle
weight are omitted from production builds.
